cleaned up file structure, eliminated redundancies, and standardized syntax
