+++ src/App.js
@@ -1,28 +1,27 @@
 import React, { Component } from 'react';
 import './App.css';
-// import Banner from "./components/Banner";
-import Candidates from "./components/Candidates";
-// import ElectionHome from "./components/ElectionHome";
-//import Issues from "./components/Issues";
-// import Login from "./components/Login";
-// import Mainpage from "./components/Mainpage";
-// import Sidebar from "./components/Sidebar";
-// import Signup1 from "./components/Signup1";
+import Candidates from "./components/Candidates/Candidates";
+import ElectionHome from "./components/ElectionHome/ElectionHome";
+import Quiz from "./components/Quiz/Quiz";
+import Login from "./components/Login/Login";
+import Mainpage from "./components/Mainpage/Mainpage";
+import Signup1 from "./components/Signup/Signup1";
+import Signup2 from "./components/Signup/Signup2";
+import Title from "./components/Title/Title";
 import pageData from "./pageData";
 
 class App extends Component {
   render() {
     return (
       <div className="App">
-
+        <Title pageData={pageData}/>
         <Candidates pageData={pageData}/>
-        {/* /* <Banner pageData={pageData}/>        
         <ElectionHome pageData={pageData}/>
-        <Issues pageData={pageData}/>
+        <Quiz pageData={pageData}/>
         <Login pageData={pageData}/>
         <Mainpage pageData={pageData}/>
-        <Sidebar pageData={pageData}/>
-        <Signup pageData={pageData}/> */ }
+        <Signup1 pageData={pageData}/> 
+        <Signup2 pageData={pageData}/>
       
       </div>
     )
