--- src/components/Banner.js
@@ -1,10 +1,0 @@
-import React, { Component } from "react";
-
-class Banner extends Component {
-    render() {
-        let pageData = this.props.pageData;
-        return ("login")
-    }
-}
-
-export default Banner;
