--- src/components/ElectionHome.js
@@ -1,10 +1,0 @@
-import React, { Component } from "react";
-
-class ElectionHome extends Component {
-    render() {
-        let pageData = this.props.pageData;
-        return ("Election home")
-    }
-}
-
-export default ElectionHome;
