--- src/components/Mainpage.js
@@ -1,10 +1,0 @@
-import React, { Component } from "react";
-
-class Mainpage extends Component {
-    render() {
-        let pageData = this.props.pageData;
-        return ("mainpage")
-    }
-}
-
-export default Mainpage;
