--- src/components/Sidebar.js
@@ -1,10 +1,0 @@
-import React, { Component } from "react";
-
-class Sidebar extends Component {
-    render() {
-        let pageData = this.props.pageData;
-        return ("sidebar")
-    }
-}
-
-export default Sidebar;
